pyproject.toml
[project]
name = "resume-jd-matching-irs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.118.0",
    "google-genai>=1.41.0",
    "httpx>=0.28.1",
    "joblib>=1.5.2",
    "nltk>=3.9.2",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.4",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "uvicorn[standard]>=0.37.0",
]
